[html-drawing] Support manual page breaks
`drawDOM` takes an additional options argument, which can contain
forcePageBreak (a CSS selector).  Elements that match this selector will
force a page break before.  Elements themselves are still included in the
output, but on the next page.  This makes it possible to, say, set a certain
class name on rows of a grid where a page break should occur.

The page break elements need not be at the toplevel -- they can be deeply
nested inside other elements.  Any parent elements up-to the one passed to
drawDOM will be split for each page break, so borders/backgrounds applied to
parent elements will be carried over to the next page.  This operation is
destructive, so it's done on a clone node which is removed from the document
when the drawing is complete.