[html-drawing] Optimize text rendering
EOL detection code took a lot of time.  Previously, we would determine where
a line ends by increasing a selection range by one character at a time,
tracking the bottom of the bounding box.  For a long line of 15K characters,
that would mean 15K calls to range.setEnd and getBoundingClientRect.

Implemented a binary search for it which reduces the number of tries to
approx log₂(line_length).

This should positively impact speed for non-huge-lines as well.

Close #4180