[spreadsheet] Third refactoring
Ditch again Promises in favor of plain callbacks.  Seems to offer the right
balance of speed and simplicity with support for async formulae (and in
theory, since the evaluator has to assume every cell could have an async
formula, we could support server-side paging too if ever needed).  I'm
finally happy with the outcome.