[spreadsheet] Fix ROWS and COLUMNS functions
"matrix" type specifier caused defineFunction to force the argument,
resulting in a CIRCULAR error when a formula like =COLUMNS(A1:C3) sits in
B2.  There's no reason for the error, since the function does not need the
cell values.

Added a "#matrix" specifier which does not try to resolve cell values.