[spreadsheet] Don't evaluate a formula twice
This fixes a stack overflow issue with mutually referencing formulas,
i.e. A1: =B1, B1: =A1.  Unfortunately, we still can't display a
"#CIRCULAR!" error in such case, because it would break async formulas:

    A1: =ASYNC(...), B1: =A1+1

Recalculating, we run A1; because it's async, we move to B1 immediately,
which goes to call A1 again.  With this patch, A1 will be in "pending"
state so we don't run it again, just store the callback for B1, and when
the result comes both A1 and B1 will update.  If we'd throw "CIRCULAR"
for a pending formula, we'd break legitimate use of async functions.