[spreadsheet] Move parse error handling where it belongs
Eventually we might want to expose `Range::_editableValue` as a public
API (with another name, perhaps `input`).  Showing an error popup there
is inappropriate.

Let this function throw the exception, and catch it in EditCommand.
Also, instead of a plain Error it is now a calc.ParseError, so we can be
specific about catching it, and it'll contain `pos` - index of the
character where the error occurred (can be used to move the caret there
in the editor).