[spreadsheet] Fix highlighting references that contain sheet name
They won't intersect with the view, which sets no sheet name on the
references it creates, and the wouldn't be displayed.  For a quick fix,
dropping the sheet name from the reference tokens.  A better fix would
be to have a sheet name in every reference object, but it seems too much
work at this point.