[spreadsheet] Small refactoring of formatting API
formatting.compile(format)(value) no longer returns a virtual DOM
element, but something like:

    { type: "date", body: [ "1/2/2003" ], color: "red" }

To remain compatible, `formatting.format` transforms that into virtual
DOM as before.  Added `formatting.text(value, format, culture)` which
returns the formatted value as plain text (should be even a bit faster
than going through the virtual DOM).

The reason for the change is we need an easier way to retrieve the text
and color for PDF support (the virtual DOM is a bit overkill for this).