Squeeze more milliseconds in XML parser
This pervert optimization minimizes calls to String.fromCharCode, which
apparently is expensive, by passing the array of bytes we're building
down to readChar and UCS2.  fromCharCode is applied just once for
building an entire string, rather than individually for each char.

It's now around 5.3s for loading the "large file".

Ref #5170