[spreadsheet] Few more layout fixes and we're one refactoring away!
Currently the issue is that a merged cell that starts *before* the range
to show, won't show up at all, even if it does in fact intersect the
view.  Contemplating two ideas:

- add special case to the `doLayout` algorithm that takes care of this
  when `forScreen` is passed

- write separate layout function for screen case

I begin to favor the second, since we clearly must take quite different
decisions when cells are to be nicely split across pages vs. just draw
on a contiguous space.