[spreadsheet] Better handling for commands on sparse range
Partially reverts 85e0ec22d7a24752aee5a1cc448593140a059b5f

This implements range.skipHiddenCells as suggested by @gyoshev -- if the
range is sparse, it returns a new UnionRef-based range, so that commands
work only on visible cells.  This seems to make sense only for
PropertyChangeCommand and descendants, so that's where we use
range.skipHiddenCells.

Ref #5523