[spreadsheet] Fix parenthesizing operators with same precedence
While parsed correctly, an expression like `1/(2*4)` was wrongfully
printed as `1/2*4`.  Due to formula caching, this had the additional
effect that when you later type `1/2*4` into a cell, you get 0.125 when
the proper result should be 2.

The fix is to include the parens when printing a `binary` node which is
the `right` side of a `binary` parent node having the same precedence.

Fix #5569