Use model field type when posting to datasource
ref #5648

I don't really like that approach, as it relies on
a lot of information from the datasource.
A datasource.fieldType(field) method would be very handy here.

An alternative approach is to provide the type information directly
through the spreadsheet (via the columns configuration).

What I like about this is that the conversion happens automatically when
the customer configures the DataSource.

/cc @rkonstantinov