Add Map Tooltip settings
The Map does not create its own tooltip settings, but instead relies on
the actual Tooltip widget to contain and serialize those.

This avoids duplicating the more complicated things such as
LoadContentFrom which are supported in the MVC 5 wrappers.

The bad part is that we need a separate template for the Tooltip
settings. Something I didn't went on to do right now. Instead I'll
disable the tooltip generation until these templates are needed and
implemented.

/cc @Stamo-Gochev