[spreadsheet] Change output style of ref.print() for absolute refs
Prefer $A$1 to R1C1 when it can be done unambiguously (that is, when the
reference is absolute).  The reason is that Excel does not understand
R1C1 notation in name definitions.

In most use cases, names should be absolute; relative refs are not
supported in Google Sheets at all, and are buggy in Excel (we do a
better job already).