Remove empty block nodes when trimming dom #6005
When trimming dom nodes (usually when splitting ranges in command executions) whitespace nodes used to be moved before the node. This breaks the order of the Dom nodes and the restore selection points could break.

In order to avoid this, keep the whitespace, but remove the block nodes that contain only whitespace nodes.