[spreadsheet] Improve INDEX function
Don't call asMatrix right away on the argument, as this will trigger
recalculation of all cells.  Not a big deal since formulas are cached
per recalculation cycle, but still, pointless when we only need to
return data in one cell/row/column.

This is for telerik/kendo#6159 -- but unfortunately, it still doesn't
fix loading the document in this issue.