refactor(spreadsheet): editor got too complex
This reverts most of adeaf684ea14cee6946bcd40469123ed94d39a5f

Following discussion, we clearly must deactivate the editor on blur.
Too many issues are caused by that.

"reject"-type validation (not working properly yet) will just display
the error in a popup, and will revert the cell to previous value,
without keeping the editor open.  The downside is that users will have
to type the value again, but this is only a minor inconvenience compared
to the plethora of issues existing currently.  Also, validation will
typically involve numbers of dates (just a bunch of characters to type
again), not long text.