fix(editor): Prevent caret from disappearing. #6370
The toolbar is opened with timeout when it overlaps the content
area in IE or Edge. This prevents the caret from disappearing
when the toolbar openes over the content area as a result of clicking
in the content area and a tool with unselectable="on" ends up in
front of the cursor. The described behavior appears to be browser
specific, because the contenat area retains the focus and the range
is not lost even if the caret is not visible and typing in the editor
is not possible.