fix: axes are not accessible on render
Closes telerik/kendo#6580

@karadachki:

The render event is triggered before the `_instance` field
[is set](https://github.com/telerik/kendo/blob/master/src/dataviz/chart/chart.js#L348)
until the constructor returns.

The fix is more of a workaround and assumes that the `_instance` field
won't change which is probably true.

A cleaner fix would be to delay the rendering of the underlying Chart so
we can do:

```
this._instance = new KendoChart(...);
this._instance.bootstrap();
```

This still looks silly, but will ensure that we have the instance
reference before it starts doing stuff.