fix(spreadsheet): include hidden data in copy/paste
Seems data should be skipped only when content is filtered (#5302); when
rows or columns are hidden manually, they should be still included in the
copy/paste.  We don't have a flag to tell us whether a row or column is
hidden as a result of filtering, or it's just manually hidden, therefore
we'll skip hidden rows/cols only when we know that the sheet is filtered.

Fixes telerik/kendo-ui-core#1808