fix(spreadsheet): fix inserting references while editing formulas
While editing formulas it is possible to switch to another sheet to pick
references using mouse or keyboard.  This commit fixes a couple of issues:

1. previously, pressing Enter would save the formula in the new sheet rather
   than the one where the editor was originally opened.

2. when we pick a reference from another sheet, it must be prefixed in the
   editor with `SheetName!...`, or else it would refer to the original
   sheet.

Also fixes a bunch of test failures that occurred before this
commit (FormulaInput element must be focused in order for setPos to work;
seems to be a new Chrome "feature").

Ref telerik/kendo-ui-core#3210