fix(spreadsheet): fix undo after insert rows/columns
Previously, the complete sheet state was saved using `sheet.getState()`, and
restored for undo.  This broken formulas, because `getState` used to not
deep-clone formulas.  Even if this is now fixed (formulas are deep-cloned in
`getState`), cross-sheet formulas will still be broken.  For example if in
`Sheet1` a formula refers to `Sheet2!A1:A3` and we insert a new row 2 in
`Sheet2`, the formula will properly update to `Sheet2!A1:A4`, but a Undo
operation will not properly restore the reference (simply because the
formula is in another sheet).

Thus, instead of relying on state, undoing an insert operation will just
apply the reverse delete operation (leading to sheet.deleteRow which updates
the affected formula references).

Part of telerik/kendo-ui-core#2816